spring-homework-08: done

+++ spring-homework-08/src/main/java/ru/otus/homework/dto/BookComments.java
@@ -3,11 +3,13 @@ package ru.otus.homework.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.Id;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class BookComments {
-    private Integer commentsCount;
-    private Integer avgCommentsRating;
+
+    private long commentsCount;
+    private double avgCommentsRating;
 }
